cookie-parser :   Backend/app.js 	changes for sending email :   Backend/config.env 	cookie implemented & sendEmail() :   Backend/controllers/admin/adminAuth.js 	npm install:   Backend/package-lock.json 	npm install:   Backend/package.json 	new file for sending emails:   Backend/utilities/email.js
