+++ Backend/config.env
@@ -1,5 +1,11 @@
 NODE_ENV=development
 PORT=5000 
 LOCAL_DATABASE=mongodb://127.0.0.1:27017/maham
+SESSIONS_DATABASE=mongodb://127.0.0.1:27017/sessions
 JWT_SECRET=MatbietRixineum 
+TOKEN_SECRET=Maham
 JWT_EXPIRES_IN=1000d
+EMAIL_USERNAME=06d3ab2351e59d
+EMAIL_PASSWORD=b580ab16e3e1fe
+EMAIL_HOST=sandbox.smtp.mailtrap.io
+EMAIL_PORT=2525
