+++ Backend/controllers/admin/adminAuth.js
@@ -1,7 +1,5 @@
 const bcrypt = require('bcryptjs');
-const Memcached = require('memcached');
 const jwt = require('jsonwebtoken');
-const nodemailer = require('nodemailer');
 const { validationResult } = require('express-validator');
 const { promisify } = require('util');
 /////////////////////////////////////////////////////
@@ -9,9 +7,8 @@ const adminDB = require('../../models/admin');
 /////////////////////////
 const catchAsync = require('./../../utilities/catchAsync');
 const AppError = require('./../../utilities/appError');
+const sendEmail = require('./../../utilities/email');
 /////////////////////////////////////////////////////
-
-const memcached = new Memcached(`localhost:${process.env.port}`);
 
 const signToken = (email, adminId) => {
 	return jwt.sign({ email, adminId }, process.env.JWT_SECRET, {
@@ -73,8 +70,6 @@ exports.signUp = catchAsync(async (req, res, next) => {
 exports.logIn = catchAsync(async (req, res, next) => {
 	const { password, inputVerificationCode } = req.body;
 	const email = req.body.username;
-	// const verificationCode = req.session.verificationCode;
-	// console.log(verificationCode);
 
 	// 1) check if email or password provided
 	if (!email || !password) {
@@ -88,18 +83,6 @@ exports.logIn = catchAsync(async (req, res, next) => {
 		return next(new AppError('admin not found', 400));
 	}
 
-	// Retrieving the token
-	memcached.get(`userToken:${admin._id}`, function (err, verificationCode) {
-		if (err) {
-			console.log('Error retrieving user token from Memcached:', err);
-		} else {
-			console.log(
-				'User token retrieved successfully from Memcached:',
-				verificationCode
-			);
-		}
-	});
-
 	const isEqual = await bcrypt.compare(password, admin.password);
 	if (!isEqual) {
 		return next(
@@ -107,9 +90,13 @@ exports.logIn = catchAsync(async (req, res, next) => {
 		);
 	}
 
-	if (verificationCode.toString() !== inputVerificationCode) {
-		return next(new AppError('verification code was wrong', 400));
+	///////////////////////////////////////////////////////////////
+	if (req.cookies.verifyToken !== inputVerificationCode) {
+		return next(
+			new AppError('token is not correct', 401) //not authorized
+		);
 	}
+	///////////////////////////////////////////////////////////////
 
 	// 3) if everything okay , create & send token to client
 	const token = await signToken(email, admin._id);
@@ -125,11 +112,11 @@ exports.verificationCode = async (req, res) => {
 	const email = req.body.username;
 	const password = req.body.password;
 
-	const admin = await adminDB.findOne({ email });
+	const admin = await adminDB.findOne({ email }).select('+password');
 
 	if (!admin) {
 		return res.status(401).json({
-			message: 'email wrong',
+			message: 'wrong email',
 		});
 	}
 
@@ -143,327 +130,28 @@ exports.verificationCode = async (req, res) => {
 
 	const verificationCode = Math.floor(1000 + Math.random() * 9000);
 
-	// Storing the token
-	memcached.set(
-		`userToken:${admin._id}`,
-		verificationCode,
-		3600,
-		function (err) {
-			if (err) {
-				console.log('Error storing user token in Memcached:', err);
-			} else {
-				console.log('User token stored successfully in Memcached');
-			}
-		}
-	);
-
-	const transporter = nodemailer.createTransport({
-		service: 'gmail',
-		auth: {
-			user: 'aydinzarifieaszo@gmail.com',
-			pass: 'rwtrwybhtbugnqxc',
-		},
+	////////////////////////////////////////////////////
+	res.cookie('verifyToken', verificationCode, {
+		expires: new Date(Date.now() + 180000),
+		httpOnly: true,
 	});
+	////////////////////////////////////////////////////
 
 	const mailOptions = {
-		from: 'aydinzarifieaszo@gmail.com',
-		to: email,
+		email: email,
 		subject: 'Ver',
-		html: `  
-    <div
-      style="
-        background: linear-gradient(to left, #ef4057, #ef4077);
-        min-height: 100%;
-        padding-top: 20px;
-        padding-bottom:20px;
-        padding-left: 10px;
-        padding-right: 10px;
-        align-items: center;
-      "
-    >
-      <div
-        class="container"
-        style="
-          min-height: 300px;
-          max-width: 530px;
-          margin: 0 auto;
-          background-color: #fff;
-          padding: 10px 10px 10px 10px;
-          border-radius: 10px;
-          box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-          text-align:center;
-        "
-      >
-          <img
-            class="logo"
-            src="cid:myImage"
-            style="width: 190px; height: 70px;
-             "
-            alt="Logo"
-          />
-        <div
-          class="info"
-          style="
-            margin-bottom: 20px;
-            width: 90%;
-            background-color:#ef4077;
-            color: #ffffff;
-            height: 40px;
-            font-size: 20px;
-            margin-left: 5%;
-            padding-top: 14px;
-            border-radius: 10px;
-            text-align: center;
-            padding-left: 10px;
-          "
-        >
-          Email Safety phrase
-        </div>
-        <p
-          style="
-            text-align: left;
-            font-family: URWDIN;
-            font-size: 14px;
-            line-height: 22px;
-            text-align: left;
-            color: rgba(1, 8, 30, 0.6);
-            padding: 0 10px 0 10px;
-          "
-        >
-          Dear Maham admin:
-        </p>
-        <p
-          style="
-            text-align: left;
-            font-family: URWDIN;
-            font-size: 14px;
-            line-height: 22px;
-            text-align: left;
-            color: rgba(1, 8, 30, 0.6);
-            padding: 0 10px 0 10px;
-          "
-        >
-          this is your Verification code for log in :
-        </p>
-        <button
-          style="
-            background-color: transparent;
-            color: #333;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-            font-weight: 500;
-            font-size: 30px;
-            text-align: left;
-            width: 100%;
-          "
-        >${verificationCode}
-        </button>
-        <p
-          style="
-            text-align: left;
-            font-family: URWDIN;
-            font-size: 14px;
-            line-height: 22px;
-            text-align: left;
-            color: rgba(1, 8, 30, 0.6);
-            padding: 0 10px 0 10px;
-          "
-        >
-          This code remains valid for 10 minutes. Please do not disclose it to
-          anyone (including Maham staff)
-        </p>
-
-        <div
-          class="code-container"
-          style="
-            margin-top: -10px;
-            padding: 10px 10px 0px 10px;
-            background-color: #f9f9f9;
-            border: 1px solid #ccc;
-            text-align: left;
-            border-radius: 10px;
-          "
-        >
-        <div style="margin: 0 auto; max-width: 568px; background-color: #f9f9f9;">
-        <table
-          align="center"
-          border="0"
-          cellpadding="0"
-          cellspacing="0"
-          role="presentation"
-          style="width: 100%"
-        >
-          <tbody>
-            <tr>
-              <td
-                style="
-                  direction: ltr;
-                  font-size: 0;
-                  padding: 0;
-                  padding-bottom: 24px;
-                  text-align: center;
-                "
-              >
-                <div
-                  class="m_-5146143764419861847mj-column-per-100"
-                  style="
-                    font-size: 0;
-                    text-align: left;
-                    direction: ltr;
-                    display: inline-block;
-                    vertical-align: top;
-                    width: 100%;
-                  "
-                >
-                  <table
-                    border="0"
-                    cellpadding="0"
-                    cellspacing="0"
-                    role="presentation"
-                    style="vertical-align: top"
-                    width="100%"
-                  >
-                    <tbody>
-                      <tr>
-                        <td
-                          align="left"
-                          class="m_-5146143764419861847list"
-                          style="
-                            background: #f9f9f9;
-                            font-size: 0;
-                            padding: 10px 0px;
-                            word-break: break-word;
-                          "
-                        >
-                          <table
-                            cellpadding="0"
-                            cellspacing="0"
-                            width="100%"
-                            border="0"
-                            style="
-                              color: #000;
-                              font-family: URWDIN;
-                              font-size: 13px;
-                              line-height: 22px;
-                              table-layout: auto;
-                              width: 100%;
-                              border: none;
-                            "
-                          >
-                            <tbody>
-                              <tr>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-title"
-                                >
-                                  Account:
-                                </td>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-content"
-                                >
-                                  <a
-                                    href="${email}"
-                                    target="_blank"
-                                    >aydinzarifieaszo@gmail.com</a
-                                  >
-                                </td>
-                              </tr>
-                              <tr>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-title"
-                                >
-                                  IP Address:
-                                </td>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-content"
-                                >
-                                  5.122.183.29
-                                </td>
-                              </tr>
-                              <tr>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-title"
-                                >
-                                  Login Platform:
-                                </td>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-content"
-                                >
-                                  ANDROID
-                                </td>
-                              </tr>
-                              <tr>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-title"
-                                >
-                                  Login Time:
-                                </td>
-                                <td
-                                  valign="top"
-                                  class="m_-5146143764419861847list-content"
-                                >
-                                  2021-09-06 15:18:44 (UTC+08:00)
-                                </td>
-                              </tr>
-                            </tbody>
-                          </table>
-                        </td>
-                      </tr>
-                    </tbody>
-                  </table>
-                </div>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    
-        </div>
-        <p
-          style="
-            text-align: left;
-            font-family: URWDIN;
-            font-size: 14px;
-            line-height: 22px;
-            text-align: left;
-            color: rgba(1, 8, 30, 0.6);
-            padding: 0 10px 0 10px;
-          "
-        >
-          Notice: If you did not conduct this operation, your account may be
-          compromised.  Please log in to your account and change your password or
-          freeze your account immediately
-        </p>
-      </div>
-    </div>`,
-		attachments: [
-			{
-				filename: 'Maham2.png',
-				path: './public/images/Maham2.png',
-				cid: 'myImage', // Content ID of the image
-			},
-		],
+		text: 'hello there ',
+		verificationCode,
 	};
 
-	transporter.sendMail(mailOptions, (err, info) => {
-		if (err) {
-			console.log(err);
-		} else {
-			console.log(info);
-		}
-	});
-
-	return res.status(201).json({
-		message: 'Success',
-	});
+	try {
+		await sendEmail(mailOptions);
+		return res.status(201).json({
+			message: 'Success',
+		});
+	} catch (err) {
+		console.log(err);
+	}
 };
 
 // not complete
