+++ Backend/package.json
@@ -12,8 +12,10 @@
 	"dependencies": {
 		"bcryptjs": "^2.4.3",
 		"body-parser": "^1.20.2",
+		"cookie-parser": "^1.4.6",
 		"ejs": "^3.1.9",
 		"express": "^4.18.2",
+		"express-session": "^1.17.3",
 		"express-validator": "^7.0.1",
 		"jsonwebtoken": "^9.0.0",
 		"memcached": "^2.2.2",
